modify newtask.py to not send a request when the log entry is empty

+++ newtask.py
@@ -9,6 +9,9 @@ s="http://127.0.0.1:45222"
 
 out = run(['dmenu', '-p', 'New log entry', '-lines', '0'], capture_output=True)
 
-requests.post(f'{s}/switchtask', data={'log-entry':f'{out.stdout.decode().strip()}'})
-os.remove('/tmp/timepassed')
+log_text = out.stdout.decode().strip()
+
+if log_text != "":
+    requests.post(f'{s}/switchtask', data={'log-entry':f'{out.stdout.decode().strip()}'})
+    os.remove('/tmp/timepassed')
